integrate gift copy into single digi subs hero
